+++ src/cogs/block.py
@@ -19,7 +19,7 @@ class Block(commands.Cog):
         utils.write_tf(ctx.author, ctx.guild, block_channel=channel)
         word = "yourself" if str(channel.id) in utils.load_tf(ctx.user, ctx.guild)['blocked_channels'] else "transformed"
         channel_word = "this channel" if channel == ctx.channel else channel.mention
-        await ctx.respond(f"You will now be {word} in {channel_word}!")
+        await ctx.respond(f"You will now be {word} in {channel_word}! (Use this same command to revert this)")
 
     @block_command.command(description="Block a user from interacting with you")
     async def user(self,
@@ -27,7 +27,7 @@ class Block(commands.Cog):
                    user: discord.User) -> None:
         utils.write_tf(ctx.author, ctx.guild, block_user=user)
         word = "unblocked" if user.id in utils.load_tf(ctx.user, ctx.guild)['blocked_users'] else "blocked"
-        await ctx.respond(f"{user.mention} has been {word} from interacting with you!")
+        await ctx.respond(f"{user.mention} has been {word} from interacting with you! (Use this same command to revert this)")
 
     @block_command.command(description="Set a channel category where you just wanna be yourself")
     async def category(self,
@@ -39,7 +39,7 @@ class Block(commands.Cog):
             utils.write_tf(ctx.author, ctx.guild, block_channel=channel)
         word = "yourself" if str(category.channels[0].id) in utils.load_tf(ctx.user, ctx.guild)['blocked_channels'] else "transformed"
         category_word = "this channel" if category == ctx.channel.category else category.mention
-        await ctx.respond(f"You will now be {word} in {category_word}!")
+        await ctx.respond(f"You will now be {word} in {category_word}! (Use this same command to revert this)")
 
 
 def setup(bot: discord.Bot) -> None: